Moves replies to helpers/replies
