+++ lib/helpers/index.js
@@ -5,6 +5,7 @@ const attachments = require('./attachments');
 const twilio = require('./twilio');
 const slack = require('./slack');
 const request = require('./request');
+const replies = require('./replies');
 const broadcast = require('./broadcast');
 
 // TODO: Add "Helper" postfix to prevent name confusion with npm modules like twilio.
@@ -14,5 +15,6 @@ module.exports = {
   twilio,
   slack,
   request,
+  replies,
   broadcast,
 };
