+++ lib/middleware/receive-message/bad-words.js
@@ -6,7 +6,7 @@ const helpers = require('../../helpers');
 module.exports = function rejectBadWords() {
   return (req, res, next) => {
     if (badWords.isProfane(req.inboundMessageText)) {
-      return helpers.badWords(req, res);
+      return helpers.replies.badWords(req, res);
     }
 
     return next();
