+++ lib/middleware/receive-message/campaign-closed.js
@@ -11,7 +11,7 @@ module.exports = function closedCampaign() {
 
     // Check if User's current Campaign has closed.
     if (gambitCampaigns.isClosedCampaign(req.campaign)) {
-      return helpers.campaignClosed(req, res);
+      return helpers.replies.campaignClosed(req, res);
     }
 
     return next();
