+++ lib/middleware/receive-message/campaign-continue.js
@@ -5,7 +5,7 @@ const helpers = require('../../helpers');
 module.exports = function askContinueTemplate() {
   return (req, res) => {
     if (helpers.isGambitCampaignsTemplate(req.lastOutboundTemplate)) {
-      return helpers.continueCampaign(req, res);
+      return helpers.replies.continueCampaign(req, res);
     }
 
     // If we're here, we may have a non-Campaign topic set.
@@ -13,7 +13,7 @@ module.exports = function askContinueTemplate() {
     return req.conversation.setCampaignTopic()
       // TODO: If Conversation.signupStatus is 'declined', we shouldn't keep asking to continue
       // the same Campaign -- send Ask Signup for a random Campaign instead?
-      .then(() => helpers.askContinue(req, res))
+      .then(() => helpers.replies.askContinue(req, res))
       .catch(err => helpers.sendErrorResponse(res, err));
   };
 };
