+++ lib/middleware/receive-message/campaign-keyword.js
@@ -26,9 +26,9 @@ module.exports = function getCampaignForKeyword() {
         return req.conversation.setCampaign(campaign)
           .then(() => {
             if (gambitCampaigns.isClosedCampaign(req.campaign)) {
-              return helpers.campaignClosed(req, res);
+              return helpers.replies.campaignClosed(req, res);
             }
-            return helpers.continueCampaign(req, res);
+            return helpers.replies.continueCampaign(req, res);
           });
       })
       .catch(err => helpers.sendErrorResponse(res, err));
