+++ lib/middleware/receive-message/campaign-menu.js
@@ -17,7 +17,7 @@ module.exports = function campaignMenu() {
 
         return req.conversation.setCampaign(randomCampaign);
       })
-      .then(() => helpers.askSignup(req, res))
+      .then(() => helpers.replies.askSignup(req, res))
       .catch(err => helpers.sendErrorResponse(res, err));
   };
 };
