+++ lib/middleware/receive-message/parse-ask-continue-answer.js
@@ -11,13 +11,13 @@ module.exports = function parseAskContinueResponse() {
 
     const text = req.rivescriptReplyText;
     if (helpers.isConfirmedCampaignMacro(text)) {
-      return helpers.confirmedContinue(req, res);
+      return helpers.replies.confirmedContinue(req, res);
     }
 
     if (helpers.isDeclinedCampaignMacro(text)) {
-      return helpers.declinedContinue(req, res);
+      return helpers.replies.declinedContinue(req, res);
     }
 
-    return helpers.invalidAskContinueResponse(req, res);
+    return helpers.replies.invalidAskContinueResponse(req, res);
   };
 };
