+++ lib/middleware/receive-message/parse-ask-signup-answer.js
@@ -11,15 +11,15 @@ module.exports = function parseAskSignupResponse() {
 
     const text = req.rivescriptReplyText;
     if (helpers.isConfirmedCampaignMacro(text)) {
-      return helpers.confirmedSignup(req, res);
+      return helpers.replies.confirmedSignup(req, res);
     }
 
     if (!helpers.isDeclinedCampaignMacro(text)) {
-      return helpers.invalidAskSignupResponse(req, res);
+      return helpers.replies.invalidAskSignupResponse(req, res);
     }
 
     return req.conversation.declineSignup()
-      .then(() => helpers.declinedSignup(req, res))
+      .then(() => helpers.replies.declinedSignup(req, res))
       .catch(err => helpers.sendErrorResponse(res, err));
   };
 };
