+++ lib/middleware/receive-message/support-message.js
@@ -13,7 +13,7 @@ module.exports = function postMessageToFront() {
     return front.postMessage(req.platformUserId, req.inboundMessage.text)
       .then((frontRes) => {
         logger.debug('front.postMessage', frontRes, req);
-        return helpers.noReply(req, res);
+        return helpers.replies.noReply(req, res);
       })
       .catch(err => helpers.sendErrorResponse(res, err));
   };
