+++ lib/middleware/receive-message/support-requested.js
@@ -9,7 +9,7 @@ module.exports = function requestSupport() {
     }
 
     return req.conversation.supportRequested()
-      .then(() => helpers.supportRequested(req, res))
+      .then(() => helpers.replies.supportRequested(req, res))
       .catch(err => helpers.sendErrorResponse(res, err));
   };
 };
