+++ lib/middleware/receive-message/template-crisis.js
@@ -5,7 +5,7 @@ const helpers = require('../../helpers');
 module.exports = function sendCrisisMessage() {
   return (req, res, next) => {
     if (helpers.isSendCrisisMessageMacro(req.rivescriptReplyText)) {
-      return helpers.crisisMessage(req, res);
+      return helpers.replies.crisisMessage(req, res);
     }
 
     return next();
