+++ lib/middleware/receive-message/template-info.js
@@ -5,7 +5,7 @@ const helpers = require('../../helpers');
 module.exports = function sendInfoMessage() {
   return (req, res, next) => {
     if (helpers.isSendInfoMessageMacro(req.rivescriptReplyText)) {
-      return helpers.infoMessage(req, res);
+      return helpers.replies.infoMessage(req, res);
     }
 
     return next();
