+++ lib/middleware/receive-message/template-no-campaign.js
@@ -8,6 +8,6 @@ module.exports = function sendNoCampaignMessage() {
       return next();
     }
 
-    return helpers.noCampaign(req, res);
+    return helpers.replies.noCampaign(req, res);
   };
 };
