+++ lib/middleware/receive-message/template-rivescript.js
@@ -9,7 +9,7 @@ module.exports = function sendRivescriptReply() {
     }
 
     return req.conversation.setTopic(req.rivescriptReplyTopic)
-      .then(() => helpers.rivescriptReply(req, res, req.rivescriptReplyText))
+      .then(() => helpers.replies.rivescriptReply(req, res, req.rivescriptReplyText))
       .catch(err => helpers.sendErrorResponse(res, err));
   };
 };
