+++ lib/middleware/receive-message/user-update.js
@@ -43,9 +43,9 @@ module.exports = function updateNorthstarUser() {
         logger.debug('northstar.updateUser success', { userId: req.userId, data }, req);
 
         if (statusUpdate === lessValue) {
-          return helpers.subscriptionStatusLess(req, res);
+          return helpers.replies.subscriptionStatusLess(req, res);
         } else if (statusUpdate === stopValue) {
-          return helpers.subscriptionStatusStop(req, res);
+          return helpers.replies.subscriptionStatusStop(req, res);
         }
 
         return next();
